+++ src/lz77.rs
@@ -113,9 +113,9 @@ impl Lz77Store {
             LitLen::LengthDist(length, dist) => {
                 let len_sym = get_length_symbol(length as usize);
                 self.ll_symbol.push(len_sym as u16);
-                self.d_symbol.push(get_dist_symbol(dist as u32) as u16);
-                self.ll_counts[llstart + len_sym as usize] += 1;
-                self.d_counts[dstart + get_dist_symbol(dist as u32) as usize] += 1;
+                self.d_symbol.push(get_dist_symbol(dist) as u16);
+                self.ll_counts[llstart + len_sym] += 1;
+                self.d_counts[dstart + get_dist_symbol(dist)] += 1;
             }
         }
     }
